requirements.txt
flask
langchain
ibm-watson
googletrans==4.0.0-rc1
python-dotenv
ibm-watsonx-ai
